dom/bindings/test/test_sequence_detection.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1066432
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1066432</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1066432 **/
  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({set: [['dom.expose_test_interfaces', true]]}, function() {
    var testInterfaceJS = new TestInterfaceJS();
    ok(testInterfaceJS, "got a TestInterfaceJS object");

    var nonIterableObject = { "@@iterator": 5 };
    if (typeof Symbol === "function") {
      // Make this test fail if Symbol.iterator is correctly implemented.
      // This is here to make sure this test is updated when bug 918828 lands,
      // at which point the correct code will be:
      //   var JS_HAS_SYMBOLS = typeof Symbol === "function";
      //   var std_iterator = JS_HAS_SYMBOLS ? Symbol.iterator : "@@iterator";
      //   var nonIterableObject = { [std_iterator]: 5};
      // Otherwise, fixing bug 918828 would cause this test to silently stop
      // testing what it's supposed to be testing.
      nonIterableObject[Symbol.iterator] = Array.prototype[Symbol.iterator];
    }

    try {
      testInterfaceJS.testSequenceOverload(nonIterableObject);
      ok(false, "Should have thrown in the overload case");  // see long comment above!
    } catch (e) {
      ise(e.name, "TypeError", "Should get a TypeError for the overload case");
      ok(e.message.contains("not iterable"),
         "Should have a message about being non-iterable in the overload case");
    }

    try {
      testInterfaceJS.testSequenceUnion(nonIterableObject);
      ok(false, "Should have thrown in the union case");
    } catch (e) {
      ise(e.name, "TypeError", "Should get a TypeError for the union case");
      ok(e.message.contains("not iterable"),
         "Should have a message about being non-iterable in the union case");
    }

    SimpleTest.finish();
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1066432">Mozilla Bug 1066432</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
